+++ Editor/Exceptions.cs
@@ -6,4 +6,8 @@ namespace VRCQuestifyer {
     public class MaterialNotFoundException : System.Exception {
         public string pathToMaterial;
     }
+
+    public class GameObjectNotFoundException : System.Exception {
+        public string pathToGameObject;
+    }
 }