+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-# 0.0.1
+# alpha2
+
+- switch material and remove game object custom actions working
+- all material fields are now copied
+- disable auto-create materials by default
+- checkbox for toon shader
+- mark action as perform at end
+- remove "copy path" extension
+- remove keyboard shortcut
+
+# alpha1
 
 - initial version